Moving ephemeral metadata impl to separate file
